Restructure the project.
