+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".view.sms.MainActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -36,18 +36,18 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".SettingsActivity"
+            android:name=".view.settings.SettingsActivity"
             android:label="@string/pref_activity_title"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".view.sms.MainActivity">
 
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value=".view.sms.MainActivity" />
         </activity>
 
         <activity
-            android:name=".ShortcutHandlerActivity"
+            android:name=".view.misc.ShortcutHandlerActivity"
             android:launchMode="singleInstance"
             android:theme="@style/Theme.AppCompat.Translucent">
             <intent-filter>
@@ -60,7 +60,7 @@
         <service android:name=".service.SendingService" />
 
         <receiver
-            android:name=".SMSStatusBroadcastReceiver"
+            android:name=".service.StatusBroadcastReceiver"
             android:enabled="true">
             <intent-filter>
                 <action android:name="com.balicodes.quicksms.SENT_STATUS"/>
@@ -69,25 +69,25 @@
         </receiver>
 
         <activity
-            android:name=".HistoryActivity"
+            android:name=".view.history.HistoryActivity"
             android:label="@string/title_activity_history"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".view.sms.MainActivity">
 
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value=".view.sms.MainActivity" />
         </activity>
 
         <activity
-            android:name=".ExportImportActivity"
+            android:name=".view.misc.ExportImportActivity"
             android:label="@string/title_activity_export_import"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".view.sms.MainActivity">
 
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value=".view.sms.MainActivity" />
         </activity>
     </application>
