+++ app/src/main/java/com/balicodes/quicksms/view/history/SendingStatusListAdapter.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.balicodes.quicksms
+package com.balicodes.quicksms.view.history
 
 import android.content.Context
 import android.graphics.Color
@@ -24,6 +24,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import android.widget.TextView
+import com.balicodes.quicksms.R
 import com.balicodes.quicksms.entity.SendStatusEntity
 import com.balicodes.quicksms.model.Status
