+++ app/src/main/java/com/balicodes/quicksms/view/sms/NoScrollListView.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.balicodes.quicksms
+package com.balicodes.quicksms.view.sms
 
 import android.content.Context
 import android.util.AttributeSet
