+++ app/src/main/java/com/balicodes/quicksms/view/sms/SMSListAdapter.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.balicodes.quicksms
+package com.balicodes.quicksms.view.sms
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -23,6 +23,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import android.widget.TextView
+import com.balicodes.quicksms.R
 import com.balicodes.quicksms.model.SMSItem
 
 internal class SMSListAdapter(private val items: List<SMSItem>) : BaseAdapter() {
