+++ app/src/main/res/layout/sms_form_fragment.xml
@@ -35,7 +35,7 @@
                     android:layout_marginTop="30dp"
                     android:hint="@string/message_hint" />
 
-                <com.balicodes.quicksms.NoScrollListView
+                <com.balicodes.quicksms.view.sms.NoScrollListView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:id="@+id/receiverListView"
